feat: add confirmation dialog service and integrate into bill deletion process

+++ src/app/app.component.ts
@@ -1,11 +1,11 @@
 import { AsyncPipe, NgClass } from '@angular/common';
 import { Component, OnInit } from '@angular/core';
-import { RouterLink } from '@angular/router';
 import { BillCreateComponent } from '@bill/bill-create/bill-create.component';
 import { DialogControl } from '@class';
-import { AccessService, ThemeService } from '@service';
+import { AccessService, ConfirmService, ThemeService } from '@service';
 import { ButtonModule } from 'primeng/button';
 import { CardModule } from 'primeng/card';
+import { ConfirmDialogModule } from 'primeng/confirmdialog';
 import { ToastModule } from 'primeng/toast';
 
 @Component({
@@ -17,10 +17,10 @@ import { ToastModule } from 'primeng/toast';
 		ButtonModule,
 		CardModule,
 		ToastModule,
+		ConfirmDialogModule,
 		NgClass,
 		BillCreateComponent,
 		AsyncPipe,
-		RouterLink,
 	],
 })
 export class AppComponent implements OnInit {
@@ -32,7 +32,8 @@ export class AppComponent implements OnInit {
 
 	constructor(
 		private themeService: ThemeService,
-		private accessService: AccessService
+		private accessService: AccessService,
+		private confirmService: ConfirmService
 	) {}
 
 	ngOnInit(): void {
@@ -46,11 +47,18 @@ export class AppComponent implements OnInit {
 	}
 
 	async deleteBill(id: string) {
-		try {
-			const response = await this.accessService.deleteDoc('bills', id);
-			console.log(response);
-		} catch (e) {
-			console.log(e);
-		}
+		this.confirmService.confirm({
+			type: 'delete',
+			option: {
+				accept: async () => {
+					try {
+						const response = await this.accessService.deleteDoc('bills', id);
+						console.log(response);
+					} catch (e) {
+						console.log(e);
+					}
+				},
+			},
+		});
 	}
 }
